Test: Results of Current DDPM Acceleration
This commit documents the results of the current implementation for
accelerating DDPM using an initial prior provided by DIP. In this
approach, the output from DIP, starting from pure noise, is used as
Gaussian noise in the forward diffusion process of the DDPM algorithm.
The evaluation process utilizes the DIP output to generate images and
produce a rate-distortion vs. timesteps plot. The implementation could
be seen in both `main.ipynb` and `denoising_diffusion_probabilistic.py`.

The test results, with 5 epochs of DIP training and 800 diffusion steps
for DDPM, are available in the output_img/800T_dip5 directory.
Unfortunately, this approach has proven ineffective, resulting only in
white images. Therefore, a more effective strategy needs to be devised
to address this issue.