Modify: The alternate approach for the task
In this commit, I explore an alternative approach to accelerate DDPM
using the initial prior from DIP.

The key difference in this approach is that instead of replacing the
Gaussian mask in the forward function with the DIP output, the DIP
output is used as target noise for the loss function calculation. The
specific changes can be seen in the code by using git diff.

For this experiment, I trained the DIP model for 1000 epochs and used
800 diffusion steps for the DDPM model. Other settings remain the same
as in the tests for the pure DDPM model.

The resulting images from this experiment are located in the
`output_img/800T_dip1000_new directory`, with the suffix new indicating
the results of this approach. Unfortunately, this method also did not
perform well in a short training period such as 100 epochs. The output
images remain noisy and are not as well-produced as those in the
`output_img/1000T` directory.

Additionally, the rate-distortion vs. timesteps plot for the current
DDPM structure, saved in the `output_img/rate-distortion_timestep_plot`
directory.

The plot reveals that the RMSE between timesteps performs differently
compared to the pure DDPM test, displaying an opposite trend, which is
quite surprising. Moreover, the RMSE across different timesteps is
significantly higher than in the pure DDPM test plot. These quantitative
metrics indicate that this approach does not perform as expected.